refactor: move zIndices to theme
