+++ src/theme/zIndices.d.ts
@@ -1,0 +1,11 @@
+type Level =
+  | `background`
+  | `base`
+  | `dropdowns`
+  | `toasts`
+  | `modals`
+  | `a11yIndicators`
+
+const zIndices: Record<Level, number>
+
+export = zIndices
