+++ src/theme/zIndices.js
@@ -1,4 +1,4 @@
-export const zIndices = {
+const zIndices = {
   background: 0,
   base: 1,
   dropdowns: 10,
@@ -6,3 +6,5 @@ export const zIndices = {
   modals: 1000,
   a11yIndicators: 10000,
 }
+
+export default zIndices
