+++ src/utils/presets/index.ts
@@ -1,3 +1,2 @@
 export { fontFamilies } from "./typography"
-export { zIndices } from "./zIndices"
 export { transition } from "./transition"
